feat: move `usePreviewMode` to `/preview`
