+++ packages/hydrogen-sanity/package.json
@@ -29,6 +29,11 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
+    "./preview": {
+      "source": "./src/preview/index.ts",
+      "import": "./dist/preview/index.js",
+      "default": "./dist/preview/index.js"
+    },
     "./preview/route": {
       "source": "./src/preview/route.ts",
       "import": "./dist/preview/route.js",
@@ -56,6 +61,9 @@
   "types": "./dist/index.d.ts",
   "typesVersions": {
     "*": {
+      "preview": [
+        "./dist/preview/index.d.ts"
+      ],
       "preview/route": [
         "./dist/preview/route.d.ts"
       ],
